Common operation module cleanup and rename
