+++ common/proxies/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2021"
 [dependencies.multiversx-sc]
 version = "=0.56.1"
 
-[dependencies.operation]
-path = "../operation"
+[dependencies.structs]
+path = "../structs"
 
 [dependencies.tx-batch-module]
 path = "../tx-batch-module"
