+++ common/tx-batch-module/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2018"
 [dependencies.multiversx-sc]
 version = "=0.56.1"
 
-[dependencies.operation]
-path = "../operation"
+[dependencies.structs]
+path = "../structs"
 
 [dev-dependencies.multiversx-sc-scenario]
 version = "=0.56.1"
