+++ sov-esdt-safe/Cargo.toml
@@ -23,8 +23,8 @@ path = "../common/utils"
 [dependencies.fee-market]
 path = "../fee-market"
 
-[dependencies.operation]
-path = "../common/operation"
+[dependencies.structs]
+path = "../common/structs"
 
 [dependencies.cross-chain]
 path = "../common/cross-chain"
