+++ sov-esdt-safe/src/deposit.rs
@@ -1,5 +1,5 @@
 multiversx_sc::imports!();
-use operation::{aliases::OptionalValueTransferDataTuple, OperationData, TransferData};
+use structs::liases::OptionalValueTransferDataTuple, OperationData, TransferData};
 
 #[multiversx_sc::module]
 pub trait DepositModule:
